Add a wait for creating crypt block device
Creating the crypt block device takes some time,
especially on low performance platform.
Currently vold isn't aware of it so sometimes vold
tries to access it before the creation is completed.
In that case the encryption or decryption would fail.
To avoid this problem, the wait for creating the block
device is added.

Change-Id: I3dd24ffc0257d5b41505cb5c27f8362aea10ec14

Conflicts:
	cryptfs.c